задача 25, возведение числа в степень

+++ task25/Program.cs
@@ -4,7 +4,7 @@ int a = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine("Введите второе число: ");
 int z = Convert.ToInt32(Console.ReadLine());
 
-int product = 1;
+double product = 1;
 for(int i = 1; i <= z; i ++)
     product = product * a;
